chainspec: use pointers for parity chainspec type
Enables unset values in geth to actually be unset in Parity.

In addition, this fixes the set-max-code-size-from-0 issue
that go-ethereum has (and multi-geth does not), and fixes the tests
so that Parity does not wrongly have a MaxCodeSizeTransition: 0x0,
instead using the proper value as set in the geth config (23000).