<many>: refactor params/types.ChainConfig to embed goetherum type
Multigeth supports all of ethereum/go-ethereum chain config data types,
and more. As multigeth grows, it will be important to keep, both for
utility and reference, a structured understanding of 'upstream' data
types.

Go's embedded struct is a perfect feature for this, so
the ethereum/go-ethereum params/ChainConfig type has been refactored
to be an embedded struct in params/types/ChainConfig.

This will enable graceful conversions between the growing
multigeth config type and go-ethereum's; this logic
is frequently wanted, eg. which matching a set of features to
a given meta fork feature-set.