cmd/geth,consensus/misc,core: remove contra-dao fork extra data validation
- Remove test for this logic
- Remove the logic
- Fix the test (want daoBlock -> want nil)

Left in comments and commented code so far for reference,
to show the changed logic.

The reasoning is that the state manipulation already decides
if a chain supports the dao fork or not. As far as I can tell,
the header extradata validation was a double-down extra-sureity
check that dao-fork-supporters would choose the correct chain.

Without extra data validation, clients on either side of the for
k should still select the proper chains because the state roots
will be different.

Signed-off-by: meows <b5c6@protonmail.com>