params/types/multigethv0: fix failing tests by setting set eip2384 -> noop
The difficulty test suite Ropsten config has EIP2384 (Muir Glacier)
set. multigethv0 doesn't support this configuration, which adjusts
difficulty, and when SetEIP2384 was called on the converted-to multigethv0
config, a Fatal error was thrown and the resulting config was invalid.

This fix simply has the config return a Noop error instead of a Fatal
one. The tests DO NOT COVER the Muir Glacier cases, though.

Resolves https://github.com/etclabscore/multi-geth/issues/139

Signed-off-by: meows <b5c6@protonmail.com>