rawdb: 'freeze' becomes a more generic function
'freeze' function was implemented as a method of
freezer instances, where freezers should be objects
that satisfy the AncientStore interface defined in
ethdb/database.go

Implementing the experimental freezerRemote object
adjacent to freezer object demonstrated the brittle
way the actual freezing logic is implented, where
that logic is responsible for transitioning data from
the KV leveldb instance to the freezer.

This change allows the 'freeze' logic to be shared
between the (now) two freezer implementations.

Signed-off-by: meows <b5c6@protonmail.com>