rawdb: use batching for s3 requests
This improves performance by remediating the one-by-one
uploads which made no use of the Sync method, which
is the method intended to do the heavy lifting for
data persistence.

Batching is managed for delete (truncate) and upload
(append) methods.

Concurrency is manually set to 10 for upload (with 5
being the default) as a shot-in-the-dark effort to get
it to speed up even more.

Right now I see about 6-8 blocks per second uploads,
with batches being 2048 blocks during a from-scratch
fast sync.

Compared to FS persistence this is still WILDLY slow,
taking around 5 minutes to sync increments of 2048 blocks.

Signed-off-by: meows <b5c6@protonmail.com>