coregeth,goethereum,multigethv0,parity: fix Clique consensus issue caused by misconfigured EIP2
EthashEIP2F interface method is poorly named.
EIP2 is NOT Ethash-specific. It is a messy EIP,
and specifies a lot of different things.
Among them, it specifies gas cost of contract creation,
changing the original cost of 21000 to 53000.

The issue cited below sees a variance of 32000 in
gas cost outcomes for a contract creation. This is
explained by the difference in the intrinsic gas cost
without and with this EIP.

Fixes https://github.com/etclabscore/core-geth/issues/120

Signed-off-by: meows <b5c6@protonmail.com>