rawdb: rename ExternalRemoteFreezer->FreezerRemoteClient
This was the name used by the original New_ func,
and I think is the most descriptive.

The business logic here is providing an RPC client
in lieu of file r/w business. Although this client
will communicate with an external server, I think its
name is best fit to Client rather than the relationship
it (the client) will have with an External server.

Signed-off-by: meows <b5c6@protonmail.com>