main: remove FreezerRemoteAPI entirely
This was a wrapper around the actual business logic
which only handled hexutil en/decode occasionally.

I believe this codec logic can be pushed to existing
RPC library implementation... and that the freezerRemoteS3
struct will or can do the job of providing the necessary
public methods (and possibly codec if that need arises).

Signed-off-by: meows <b5c6@protonmail.com>