main,server: striking for DRYer simplicity
We're already importing a lot of stuff from go-ethereum
in this cmd/ancientstore program, a program which I understand
is a sort of mock-up of what would eventually be an entirely
stand-alone program (a server) that handles business logic
as a server and as a storage provider.

Since many vendored deps are already assumed, eg
flags, rpc.Server, rpc.API, node, ...
It seems to me we might as well (for now) go
whole-hog and try to piggy-back as much existing server
code as we can.

So I struct most of the server/server.go (all of it?)
and replaced it with some out-of-the-box rpc.Server
initialization logics for http and ipc.

I'm not sure that the codec (ie hexutil) will work as expected
(this may have broken that), and I don't see a good
existing way of testing that yet at all.

If it is broken, we can add the codec type at the freezerS3 method
return values (typings), or use the rpc.ServeCodec method to
do custom things.

Signed-off-by: meows <b5c6@protonmail.com>