main: finish refactoring cacheing to use lru
This completes the trashing of the handrolled
cache and replaces it with the lru cache.

- Cache sizes are configured to be a little
overkill. This won't hurt anything, since their
limits are never really used (ermm.. their
lru-ness is never really used...)
but just provides enough space for the
magic to happen. Blocks are Remove d
once they have been pushed in a complete batch.

This completes the work begun with 86f3dbc51b.

Signed-off-by: meows <b5c6@protonmail.com>