rawdb: move ancient level query to for loop, manually increment guage
numFrozen replaces f.frozen (since the iterface doesn't provide
a field), replacing it with a function call to frdb.Ancients,
which provides the same value.

However, since we're not accessing the field directly, we need to
be careful about how we handle the logic around the assigned
value. The most implementation most equivalent to prior behavior
would be be call the method every time in lieu of f.frozen.
But this would be a lot of calls, and probably quite a bit slower.

The value of numFrozen will increment +1 every time AppendAncient
succeeds, so that logic is added. And rather than assigning the
value once prior to the for loop, the value is assigned with every
iteration of the loop via the method Ancients() call.

backoff logic is expected to handle method thrashing in the
same way that it wants to handle avoiding db thrashing.

Signed-off-by: meows <b5c6@protonmail.com>