main,utils,rawdb,eth,node,rpc: refactor --ancient.datadir.remote.endpoint -> --ancient.rpc
- Remove IPC vs. HTTP/WS transport segregation, along with
node.Config.__IPC bool. We originally segregated IPC thinking
it would reduce ambiguity in configuration, but at this point
I think it causes an unworthy amount of finnicky and adhoc
logic through the configuration and constructor codes.

This change uses the exiting rpc.Client.Dial implementation
as the reference implementation, and simplifies the remote ancient
flags to one: --ancient.rpc. On the backend, then, this flag
value is simply passed to that rpc method. The node package
validates the URL with copy-pasted code from the RPC package
when necessary (for handling an empty-value default... which
arguably it shouldn't be doing at all anyways.)

Signed-off-by: meows <b5c6@protonmail.com>