node: remove freezerURL checks/mutations from node Open- method
This logic was checking and mutating a freezer url in
a way that was a vestige of how the freezer path is handled
for the builtin freezer fs implementation.

Instead of using supervisory logic, we can just let the raw
value pass through to the rpc package where it will be
validated and used in context. Any errors coming from that
context will be raised.

This change also includes a tidbit of a wider parameter renaming
that I'm going to commit next.

Signed-off-by: meows <b5c6@protonmail.com>