cmd/ancient-store-fs: rm -rf cmd/ancient-store-fs: out of scope
This was a sketch of a good idea, but is out of scope
for the current feature.

The deleted code established a remote freezer server
that would behave exactly like the built in filesystem
freezer, but do so via RPC communication, and thus
be able to replicate geth's ancient datadir store
remotely and possibly for multiple clients.

But not yet.

Signed-off-by: meows <b5c6@protonmail.com>