go.mod,node: clean up eth_subscribe options schema def
eth_subscribe's second parameter is optional and
depends on the first parameter, which is
tricky to completely represent in json schema/openrpc
land.

This simplifies the original attempt at doing so,
using a plain ol empty interface (and fixing up
the management of THAT along the way).

This commit bumps the version of go-openrpc-reflect,
which also (now) better handles schemas with
boolean additionalProperties annotation values.

Date: 2020-11-11 16:37:42-06:00
Signed-off-by: meows <b5c6@protonmail.com>