eth,ethapi: implement typed header/+block responses for RPC methods
This is an initial plug-in of the typings for Header
and Block responses for relevant RPC methods.

THE API RESULTS ARE UNTESTED.
The source will build and return a successful
and much more descriptive response to rpc.discover,
but the equivalence and even basic functioning of
the response for, say, eth_getBlockByNumber is
untested.

Date: 2020-11-14 08:11:31-06:00
Signed-off-by: meows <b5c6@protonmail.com>