ethclient,node: add inproc rpc.discover handling, plus test
This adds a basic long-overdue test for
basic functionality of rpc.discover method
via the inproc handler.

This also adds rpc.discover support for the inproc
handler, which was necessary for the test to pass.

The number of methods is explicity hardcoded
and tested. Although this is somewhat brittle,
the hope is that it is also sensitive to
undesireable or unexpected variations.

Date: 2020-11-16 16:11:53-06:00
Signed-off-by: meows <b5c6@protonmail.com>