main: remove fakepow.poisson flag from 'copydb' command
Since this flag is only useful for mining (block
emission), we don't need it for copydb, which just
runs the import of chain data.

Fakepow can be used to import more quickly
by cutting corners and not doing ethash field
verification, eg if you really trust the chaindata
that you're copying.

Date: 2020-12-04 09:10:42-06:00
Signed-off-by: meows <b5c6@protonmail.com>