rawdb: ignore ancient writes below frozen marker
This issue has becomee very annoying for me.
https://github.com/etclabscore/core-geth/issues/245

This resolves it, although obviously not in an ideal way.
There has got to be an off-by-one somewhere that, when
the chain get 'Rewind to 1', or somewhere in download->import,
block #1 is attempted for ancient storage duplicitously.

The code comments hopefully explain why its OK
to ignore AppendAncient when the block is below
the ancient water level.

Date: 2020-12-11 10:45:45-06:00
Signed-off-by: meows <b5c6@protonmail.com>