main: init --syncmode.fast and --datadir flags
I'm having trouble keeping the light client
connected. I don't see why I shouldn't be able to
use a fast/full client for the faucet if I want.
This adds --syncmode.fast boolean flag to the faucet
program to support this.

Also adds a --datadir flag because I'd like to be able
to specify that.

Date: 2020-12-21 07:25:46-06:00
Signed-off-by: meows <b5c6@protonmail.com>