Feat/ecip1100 nodisable (#304)
It uses the flag --ecbp1100.nodisable, a bool flag that, if set, is by default truthy, and prevents the efficacy of auto-shutoff logic, but does not force activation, of ECBP1100. That is, the node will start and sync following existing behavior without ECBP1100 enabled, but once achieving sync with a best peer and then enabling ECBP1100, shutoff mechanisms based on minimum peers or stale heads will be inoperative.

* core: rename bc AF field appending _Status

Hopefully this makes the fields value a little
clearer (as a status/state rather than setting).

Date: 2021-01-14 12:05:02-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* main,cmd/utils,core,eth: init implement configurable ECBP1100 force-on override

This provides a CLI flag --ecbp1100.nodisable which
forces the ECBP1100 logic to an always state, prevent
the auto-shutoff mechanisms (trigger on below-threshold peer count
and stale head for a time beyond a certain time window) from
disabling the ECBP1100 artificial finality logic.

Date: 2021-01-14 12:07:26-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* main,cmd/utils,core,eth: refactor ecbp100 force config to on/off states

Date: 2021-01-14 12:24:55-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* core,eth: add blockchain method for checking AF override state

Comments are added noting that
- the protocol manager 'safety loop' is short-circuited
  if the state is force-set, only for the reason of saving
  a noop goroutine. Note that if the override becomes
  configurable at runtime, this would need to be cut.
- min peers check is allowed to still call the EnableArtificialFunction(true/false)
  since it will be a noop either way if the override is in place.

Date: 2021-01-14 12:40:15-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* main,cmd/utils: rename variable to better meet use

Date: 2021-01-14 12:53:21-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* core: fix assignment to nil by initializing nonzero placeholder

Date: 2021-01-14 12:59:43-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* main: fix formatting issue for crit log

Date: 2021-01-14 12:59:55-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* eth: remove optimization short-circuiting safety loop

The cost of one noop goroutine vs. potential
unexpected logic is not worth it. Would rather
call the noop and keep the AF enable/disable
logic in one place.

Date: 2021-01-14 13:06:12-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* cmd/geth,cmd/utils: change flag from string on|off to bool

Now use becomes --ecbp1100.force

where that enables forceful activation.

Date: 2021-01-15 05:19:35-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* cmd/geth,cmd/utils,core,eth: --ecbp100.nodisable deactivates auto-shutoff mechanisms

Modifies preceding implementation by changing logic
from force-always-on to permanent-once-activated state
overrides.

This allows the node to start and sync without ECBP1100
activated, but once synced with a best peer and (following
incumbent logic) having activated ECBP1100, prevents the auto-shutoff
mechanisms from having any effect;
once ECBP1100 is enabled, it will stay enabled.

Date: 2021-01-15 05:40:46-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* eth: add test for AF nodisable switch

Shows that AF is not disabled when node 'a's
peer minimum is raised back to 5, and having
only one peers.

Date: 2021-01-18 06:08:35-06:00
Signed-off-by: meows <b5c6@protonmail.com>

* typo

Co-authored-by: Chris Ziogas <ziogaschr@gmail.com>