eth: fix tests to handle stale interval realistically
Tests prior assumed the default generated chain values
which situate themselves from nearly 1970, resulting
in apparently very old chains.

This solution implements a blockgen function which
offsets the first block generated so that it is
nearly contemporary, placing it behind the current
time at an offset sufficient to fit the anticipated
generated blockchain.

With this, the tests can use the stale head
safety condition realistically since the test
chain's current head will be close enough
to the real current time.

Date: 2021-02-22 08:42:44-06:00
Signed-off-by: meows <b5c6@protonmail.com>