consensus/clique,core,params/types/ctypes: init implement Clique EIP3436 logic
This writes an initial pass of EIP3436 logic
as well as installs it in the block-writer
method of the blockchain insertion program.

In general, this points to an awkward inherited
structure of the chain-insertion program defining
and handling competing-block consensus rules.

Ideally I think the consensus Engine type
should implement an interface method like
PrioritizeHeaders, which would be intended
to handle the logic of whether or not a
chain segment should be reorganized in favor
of another.

This patch does not include tests, and is provided
only as a design sketch and general sanity check.

Date: 2021-04-16 09:02:54-05:00
Signed-off-by: meows <b5c6@protonmail.com>