cmd/geth,consensus/clique,consensus,consensus/ethash,core: rename method IsReorg -> ElectCanonical, &vars, +comments
Renaming IsReorg -> ElectCanonical addresses the fact
that although the original logic and names used in this
control flow used 'reorg', the logic was actually deciding
if an incoming block should be eligible to be the new
chain head (marked as canonical head).

Reorgs are only necessary if the incoming block IS
eligible to be canonical and DOES NOT have the current
canonical block head as its parent.

I've added and adjusted some comment around this flow
to hopefully make it better explained and readable.

Date: 2021-04-16 12:08:21-05:00
Signed-off-by: meows <b5c6@protonmail.com>