core: AMESS+: add weighting scheme to prioritize first part of chain segments
This implements a draft idea which recognizes that
for an adversary, secrecy is of greater importance for
the FIRST part of malicious segments compared to later
part(s); this is a continuous-curve kind of idea.

We can use this characteristic to emphasize (weight)
the oldest blocks more than we do the later blocks,
and in doing so, make attacks that demand initial
secrecy more risky.

Date: 2021-07-09 12:24:41-05:00
Signed-off-by: meows <b5c6@protonmail.com>