params/types/ctypes: count difficulty bomb delay implementation more leniently
Before this, the logic demanded that a difficulty bomb
delay value be matched exactly to yield an affirmative
result (indicating the EIP criteria had been met).
With this logic, the Berlin config was failing because
it implemented EIP1234 and EIP2384, where 2384 was 9m
and 1234 was 5m.

This fixes it so if the total difficulty bomb delay value
tallied from the map meets OR EXCEEDS the value spec'd by
the given difficulty bomb delay EIP, then that EIP spec
is considered implemented.

Date: 2021-11-16 10:39:11-08:00
Signed-off-by: meows <b5c6@protonmail.com>