cmd/ancient-store-mem/lib,core/rawdb: fix 'exit status 1' error on remote freezer tx indices test
a. The method was not implemented properly
because the caller (the 'client') was not
filling the argument values.
b. The mock server was trying an impossible
truncation on the maxBytes. This is a useless
error because we dont actively support the freezer
server implementation beyond the mock memory
type. So I've just removed the maxBytes feature
and it seems to work OK. Also the freezer
doesn't expect compression, so we dont have
to navigate that hefty logic like they do
with the FreezerTable.RetrieveItems.

Date: 2022-01-24 14:53:17-08:00
Signed-off-by: meows <b5c6@protonmail.com>