accounts,cmd/utils: implement coinType global configuration with defaults
A setter SetCoinTypeConfiguration is defined
to configure global vars and defaults given some
coin type.

The default is configured for Ethereum mainnet via
an init() function call in the hd.go file.

Configuration for testnet and Ethereum Classic paths
are supported via flags.go which uses this
setter when the flag --usb is used in conjunction
with a supported chain flag, eg. --ropsten (any testnet)
or --classic.

Date: 2022-07-06 10:28:46-07:00
Signed-off-by: meows <b5c6@protonmail.com>