consensus/ethash: test that ethash cache gets return expected epochs
This test reproduces an error encountered on
the Mordor testnet and reported here
https://github.com/etclabscore/core-geth/issues/439

Ethash returns a cache value with epoch
length of 30k, where it should return one
appropriate for the ECIP1099-era with
an epoch length of 60k.

This caused header verification to fail
because the mix digest was incorrect.

This issue only occurred while syncing the Mordor
network because the cache was in-memory and
because the original-era epoch was matched
later by the ECIP1099-era epoch and the LRU
future item was anachronistic.

Date: 2022-08-30 08:35:32-07:00
Signed-off-by: meows <b5c6@protonmail.com>