tests: remove dead code (state,diff mgen artifacts)
This code was used in earlier version of
the test generation logic, which extended
some given xclient suite with ETC support
by writing the results of a test to a new file.

Difficulty tests at the time used NDJSON
as an additional encoding, but this is no longer.

Now, core-geth uses xclient tests generated by
retesteth (which uses core-geth as a filler client;
actually the t8ntool command provided by the repo).

Resolves

- https://github.com/etclabscore/core-geth/pull/510#discussion_r1052683037
- https://github.com/etclabscore/core-geth/pull/510#discussion_r1052683214

Date: 2022-12-20 06:28:24-08:00
Signed-off-by: meows <b5c6@protonmail.com>