.github/workflows: .github/workflows: .github/workflows: .github/workflows: go-generate-check changed logic for whitelisting changes in generated files directly
The intention of this commit is to not break/change the logic for developers when coding.
This script serves as a warning notification if a developer forgets to generate those files using gencodec.

On specific cases where generated files are modified, we revert the modifications and compare the files so as they will show up any changes in the original files.

The developer then will find out the correct way to resolve generating the new files and keeping the modifications intact if needed.