consensus/misc,core,miner,params/types/coregeth,params/types/ctypes,params/types/goethereum,params/types/internal,params/types/multigeth,params/types/parity: configurator += Get/Set ElasticityMultiplier
This is not well supported by multi-geth or parity,
since I expect to soon drop support for them both
entirely.

The backing value for all callers is a global
var, and setters should not operate on it.
ethereum/go-ethereum ChainConfig.ElasticityMultiplier
returns a constant.
The function being a function appears to be a theoretical
placeholder.

Date: 2023-02-20 08:48:57-08:00
Signed-off-by: meows <b5c6@protonmail.com>