convert_test,params/confp/testdata,params/types/coregeth,params/types/ctypes,params/types/goethereum,params/types/internal,params/types/multigeth,params/types/parity: ElasticityMultiplier and BaseFeeChangeDenominator getter/setter no pointers
confp/TestConvert failed with a panic because the
reflection use of getters/setters
was causing a *uint64 to be passed where
the arg should have been a uint64.
Getters and setters both need to accept
and return the same type of value.

This commit also add JSON genesis dumps
for foundation,classic::go-ethereum,core-geth
configs as testdata, replacing the stureby configs.

Date: 2023-02-21 08:03:37-08:00
Signed-off-by: meows <b5c6@protonmail.com>