cmd/geth: dao test: failed to retrieve chain config
This test initializes a chain by running
geth with a temporary (eg /tmp) datadir,
then checks that the expected config can be
read from that database (before continuing
with its DAO test checks).
The config was not read back from the db.
This was because the database identifier was wrong,
so the config was written to core/geth/chaindata
instead of geth/chaindata.

Date: 2023-02-23 07:51:40-08:00
Signed-off-by: meows <b5c6@protonmail.com>