consensus/ethash: etc sync still stuck at 11700000; +TestEtchash_11700000
This test asserts that a properly-configured
Ethash will correctly validate the canonical
block at which ECIP1099 was implemented on ETC.

The test uses canonical chain data as its control
value.

Note that the test takes a long time to run;
about 2 minutes.
But considering ETC depends critically on
this function, the time seems well spent to me.

Date: 2023-02-24 06:19:39-08:00
Signed-off-by: meows <b5c6@protonmail.com>