core/vm: undefined: params.Rules
This is a tricky one. This new function is
intended for export (eg. for use with state test gen)
but core-geth does not have params.Rules, so
unfortunately we have to change the signature
to satisfy what core-geth can do. On the upside,
in order for the caller to be able to provide a
type params.Rules, they need to have the same information
(block num, time) that they would need to create
a Rules struct, so its not really breaking usability
except for the hassle of the difference.

Date: 2023-05-01 07:53:32-07:00
Signed-off-by: meows <b5c6@protonmail.com>