params: install checkpoint to config
By installing a checkpoint we'll avoid
some snap sync issues where we're attempting
to sync with very low, unsynced, nodes
and running into issues with pivoting to
full sync too early because of that (or
encountering bugs related to aborted sync rewinds).

The current block number is 18_666_111, so
the checkpoint block 18_186_239 is well
below that; safely canonical.

I just derived the checkpoint values
manually, though if we get the les client
or server working again we can theoretically
use the API les.latestCheckpoint.

Date: 2023-11-06 11:03:47-07:00
Signed-off-by: meows <b5c6@protonmail.com>