eth/fetcher,eth,miner: disable downloader-directed autopause if fetcher imports ok
The BlockFetcher handles importing announced
and broadcasted blocks, which are block notifications
supposed by peers to be at or near the canonical
head.

If we successfully import a block by the fetcher
channel, then, we can suppose ourselves to be
synced, and therefore should not longer
handle downloader sync attempts as eligible
for pausing mining.

This design assumes that the fetcher's
parameters assure that we can only import
relevant and valid head blocks locally;
ie. -7<n<32 blocks from our local head,
and having valid headers with existing parents,
etc.

Date: 2023-11-28 08:24:14-07:00
Signed-off-by: meows <b5c6@protonmail.com>