cmd/devp2p/internal/ethtest: TestEthSuite,TestSnapSuite: configure ethash ModeFake, fixes tests
The testdata use ethash as their consensus engine,
but define 0-values for all mixhashes.
So we need to not validate the PoW seals
for the test.

See the comment about how we might create a second
testdata suite that would actually define
mixhashes which could be validated by a real ethash
engine.

Date: 2024-02-29 09:40:40-07:00
Signed-off-by: meows <b5c6@protonmail.com>