core/vm: FIX precompiled addrs -> ACL
With the make..len logic, a slice
is allocated with default addr=zero keys,
then values are appended ON TOP of those.

This is different logic than master version,
which appends to an empty slice.

This fixes a bug because we dont want
a list of zero-address n-precomps long,
which will cause the zero-address to
be WARM, which it shouldnt be.

Date: 2024-08-13 09:17:00-06:00
Signed-off-by: meows <b5c6@protonmail.com>