eth/filters: fix TestPendingLogsSubscription
Pending logs were removed in v1.14.0,
but core-geth has reverted that commit.
See 0f5494e4ad92e93ce40d100dcb94be828785580f.

In particular, the fix in this patch
is the installation of backend.pendingLogsFeed.Send(flattenLogs),
where before only backend.chainFeed was getting
a useless send of the first block.

The helper function notifyPending is removed
because it was only used in one place
and unnecessarily generated a chain.

I swap t.Fatalf for t.Errorf because
it was more useful to see which subtests failed
than it was to halt at the first failure.

Date: 2024-08-14 05:48:59-06:00
Signed-off-by: meows <b5c6@protonmail.com>